Reservoir_Upscaling_x&y_direction
The reservoir was divided into 10 layers. They were upscaled layer by layer by importing Porosity and Permeability data (txt and GRDECL format) in Matlab.
The upscaling method is flow-based by addressing pressure solution for flows within the reservoir. 
The goal of this is to load this data to ECLIPSE100 to conduct simulation for optimising RF and UR with a choice of reservoir depletion method.