Porosity Data in both X and Y direction
Porosity data is in dimension of 120x60x10. So, it is 72000x1 in txt format. This is then altered into desired 60x30x10 (18000x1 in txt format) using MATLAB. Porosity upscaling is not conducted in x and y direction separately as in permeability upscaling because porosity is static data different from permeability data.