Upscaling permeability in X-direction
This is MATLAB code to execute upscaling of permeability in X-direction. The upscaling method used is "Flow-Based Upscaling". The description of what is "Flow-Based Upscaling" is written at Read Me. The imported data is "PERMX_5 - Copy.GRDECL" in 120x60x10 dimension. The result is upscaled permeability data in 60x30x10. The resultant data is stored separately into 10 txt files such that 60x30x(1-10) since the reservoir is divided into 10 layers. Transition Images from fine-scaled to coarse-scaled permeability can be found at the report listed in this Repository.