Upscaling porosity in both X and Y direction
This is MATLAB code to execute upscaling of porosity in both X and Y direction. The upscaling method used is "Arithmetic Upscaling". The description of what is "Arithmetic Upscaling" is written at Read Me. The reason of using this method is due to porosity data being static data. The imported data is "PORO_5 - Copy.GRDECL" in 120x60x10 dimension. The result is upscaled porosity data in 60x30x10. The resultant data is stored separately into 10 txt files such that 60x30x(1-10) since the reservoir is divided into 10 layers. Transition Images from fine-scaled to coarse-scaled porosity can be found at the report listed in this Repository.